requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cloudpickle==2.0.0
dask==2021.12.0
Django==4.0
djangorestframework==3.13.1
environ==1.0
fsspec==2021.11.1
locket==0.2.1
numpy==1.21.4
packaging==21.3
pandas==1.3.5
partd==1.2.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scikit-allel==1.3.5
six==1.16.0
sqlparse==0.4.2
toolz==0.11.2
